Alerts: Added Alerts to the application
This commit brings a major UI change, introducing alerts.
1. If the Name, Email or the Password field are empty a danger alert is triggered.
2. If the Passwords don't match in the verify passwords and the passwords field
then a danger alert is triggered.
3. If the register form has no errors then a success alert is triggered.

The alert component is built using the context API which has the SET_ALERT and
a REMOVE_ALERT types.
the uuid has been implemented currently to mock the backend server. its use is to
assign a unique id to each of the alert objects.
An immutable timeout of 5 seconds has been implemented for the component.