Regiser: Added the Register user in auth state.
This commit introduces the register user functionality in the app.
This commit also introduces an alert which shows that the user is already
registered. All types have been handled in the alertReducer script.
The /api/users route is causing an unknown error which is to be resolved
in future commits.